# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enspatium API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "enspatium_session"
      }
    },
...
  },
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "createUser",
        "tags": [
          "users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "displayName"
                ],
                "properties": {
                  "email": {
                    "minLength": 1,
                    "maxLength": 320,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "maxLength": 1024,
                    "type": "string"
                  },
                  "displayName": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "displayName",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "displayName",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/tokens": {
      "post": {
        "operationId": "createPersonalAccessToken",
        "tags": [
          "tokens"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "scopes"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "scopes": {
                    "minItems": 1,
                    "maxItems": 2,
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "git:read"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "git:write"
                          ]
                        }
                      ]
                    }
                  },
                  "expiresAt": {
                    "format": "date-time",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "scopes",
                    "expiresAt",
                    "lastUsedAt",
                    "revokedAt",
                    "createdAt",
                    "token"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "scopes": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "git:read"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "git:write"
                            ]
                          }
                        ]
                      }
                    },
                    "expiresAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "lastUsedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "revokedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listPersonalAccessTokens",
        "tags": [
          "tokens"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "scopes",
                      "expiresAt",
                      "lastUsedAt",
                      "revokedAt",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "scopes": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "git:read"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "git:write"
                              ]
                            }
                          ]
                        }
                      },
                      "expiresAt": {
                        "anyOf": [
                          {
                            "format": "date-time",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "lastUsedAt": {
                        "anyOf": [
                          {
                            "format": "date-time",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "revokedAt": {
                        "anyOf": [
                          {
                            "format": "date-time",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/tokens/{tokenId}": {
      "delete": {
        "operationId": "revokePersonalAccessToken",
        "tags": [
          "tokens"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "tokenId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ]
}